Data.Generator

Description

A Generator c is a possibly-specialized container, which contains values of type Elem c, and which knows how to efficiently apply a Reducer to extract an answer.

Since a Generator is not polymorphic in its contents, it is more specialized than Data.Foldable.Foldable, and a Reducer may supply efficient left-to-right and right-to-left reduction strategies that a Generator may avail itself of.

Generators

class Generator c where

minimal definition mapReduce or mapTo

Minimal complete definition

Nothing

Associated Types

type Elem c

Methods

mapReduce :: (Reducer e m, Monoid m) => (Elem c -> e) -> c -> m

mapTo :: (Reducer e m, Monoid m) => (Elem c -> e) -> m -> c -> m

mapFrom :: (Reducer e m, Monoid m) => (Elem c -> e) -> c -> m -> m

Generator Transformers

newtype Keys c

a Generator transformer that asks only for the keys of an indexed container

Constructors

Keys

Fields

getKeys :: c

newtype Values c

a Generator transformer that asks only for the values contained in an indexed container

Constructors

Values

Fields

getValues :: c

newtype Char8 c

a Generator transformer that treats Word8 as Char This lets you use a ByteString as a Char source without going through a Monoid transformer like UTF8

Constructors

Char8

Fields

getChar8 :: c

Combinators

reduce :: (Generator c, Reducer (Elem c) m, Monoid m) => c -> m

Apply a Reducer directly to the elements of a Generator

mapReduceWith :: (Generator c, Reducer e m, Monoid m) => (m -> n) -> (Elem c -> e) -> c -> n

reduceWith :: (Generator c, Reducer (Elem c) m, Monoid m) => (m -> n) -> c -> n