Data.Semigroup.Apply

Description

Semigroups for working with Apply

Documentation

newtype Trav f

A Trav uses an glues together Applicative actions with (*>) in the manner of traverse_ from Data.Foldable. Any values returned by reduced actions are discarded.

Constructors

Trav

Fields

getTrav :: f ()

newtype App f m

A App turns any Apply wrapped around a Semigroup into a Semigroup

Constructors

App

Fields

getApp :: f m