Data.Semigroup.Generator

Description

A Generator1 c is a possibly-specialized container, which contains values of type Elem c, and which knows how to efficiently apply a Reducer to extract an answer.

Generator1 is to Generator as Foldable1 is to Foldable.

Generators

class Generator c => Generator1 c where

minimal definition mapReduce1 or mapTo1

Minimal complete definition

mapReduce1

Methods

mapReduce1 :: Reducer e m => (Elem c -> e) -> c -> m

mapTo1 :: Reducer e m => (Elem c -> e) -> m -> c -> m

mapFrom1 :: Reducer e m => (Elem c -> e) -> c -> m -> m

Combinators

reduce1 :: (Generator1 c, Reducer (Elem c) m) => c -> m

Apply a Reducer directly to the elements of a Generator

mapReduceWith1 :: (Generator1 c, Reducer e m) => (m -> n) -> (Elem c -> e) -> c -> n

reduceWith1 :: (Generator1 c, Reducer (Elem c) m) => (m -> n) -> c -> n